requirements.txt
Flask==3.0.3
requests==2.32.3
gunicorn==22.0.0
Werkzeug==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
openai==1.35.3
PyPDF2==3.0.1
python-docx==1.1.2
bleach==6.1.0

